Updated how the hooks value is stored
Updated firebase exports
