+++ src/hooks/Auth/AuthProvider.jsx
@@ -1,40 +1,8 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
 import AuthContext from './AuthContext';
-import { auth, provider } from '../../firebase';
-
-const useAuthProvider = () => {
-	const [user, setUser] = useState(null);
-
-	const signIn = () => auth
-		.signInWithPopup(provider)
-		.then(({ user: authUser }) => setUser(authUser));
-
-	const signOut = () => auth
-		.signOut()
-		.then(() => setUser(false));
-
-	useEffect(() => {
-		const unsubscribe = auth
-			.onAuthStateChanged((authUser) => {
-				if (authUser) {
-					setUser(authUser);
-				} else {
-					setUser(false);
-				}
-			});
-
-		return () => unsubscribe();
-	}, []);
-
-	return {
-		isSignedIn: Boolean(user),
-		signIn,
-		signOut,
-		user,
-	};
-};
+import useAuthValue from './useAuthValue';
 
 const propTypes = {
 	children: PropTypes.oneOfType([
@@ -44,7 +12,7 @@ const propTypes = {
 };
 
 const AuthProvider = ({ children }) => {
-	const value = useAuthProvider();
+	const value = useAuthValue();
 	return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
 };
