+++ src/reducer/index.js
@@ -1,16 +1,4 @@
-export const initialState = {
-	containerTitle: null,
-	search: {
-		show: false,
-		text: '',
-	},
-	settings: {
-		darkTheme: JSON.parse(localStorage.getItem('settings-darkTheme')) || false,
-		sortNotes: localStorage.getItem('settings-sortNotes') || 'date-asc',
-		sortNotesFavourite: JSON.parse(localStorage.getItem('settings-sortNotesFavourite')) || false,
-	},
-	updateAvailable: false,
-};
+export { default as initialState } from './initial-state';
 
 export const reducer = (state, action) => {
 	const [location, name] = action.type.split('-');
